Create README.md
When changing the size of my image using width, it doesn't change the size. For example under the class wc-top-left, and changing width to 80% it is still taking
up 100% of the page and im not sure what the resolve is. Also am wondering how i can make it so that my images go underneath one another, for example i want my
image under class wc-left-side to be below wc-top-left